user-service/requirements.txt

Flask==2.2.5
PyJWT==2.3.0
Werkzeug==2.2.2
redis==4.5.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
psycopg2 ==2.9.5